msm: camera: CSID virtualization
This mechanism allows to take control over camera data path
in order to limit access to the camera frame streams. If
enabled the CSID address space is exclusively available to TEE.
TEE is responsible to actually read/write the CSID registers
according to the information relayed from the HLOS driver.

Change-Id: Ifae30241b704d8d9bc265ebff96a35c6266f2911
Signed-off-by: Shadul Shaikh <sshadu@codeaurora.org>